move all core parsers tests in test/core/parsers dir
