+++ test/core/parsers/api_query_test.js
@@ -1,33 +1,19 @@
 /**
  * Test: Parser apiParam
  */
-const assert = require('assert');
 
 // lib modules
-const parser = require('../../lib/core/parsers/api_body');
+const parser = require('../../../lib/core/parsers/api_query');
+const assert = require('assert');
 
-describe('Parser: apiBody', function () {
+describe('Parser: apiQuery', function () {
   // TODO: Add 1.000 more possible cases ;-)
   const testCases = [
-    {
-      title: 'Simple fieldname only',
-      content: 'simple',
-      expected: {
-        group: 'Body',
-        type: undefined,
-        size: undefined,
-        allowedValues: undefined,
-        optional: false,
-        field: 'simple',
-        defaultValue: undefined,
-        description: '',
-      },
-    },
     {
       title: 'Type, Fieldname, Description',
       content: '{String} name The users name.',
       expected: {
-        group: 'Body',
+        group: 'Query',
         type: 'String',
         size: undefined,
         allowedValues: undefined,
@@ -41,7 +27,7 @@ describe('Parser: apiBody', function () {
       title: 'Type, Fieldname, Description',
       content: '{String|String[]} name The users name.',
       expected: {
-        group: 'Body',
+        group: 'Query',
         type: 'String|String[]',
         size: undefined,
         allowedValues: undefined,
@@ -55,7 +41,7 @@ describe('Parser: apiBody', function () {
       title: '$Simple fieldname only',
       content: '$simple',
       expected: {
-        group: 'Body',
+        group: 'Query',
         type: undefined,
         size: undefined,
         allowedValues: undefined,
@@ -77,36 +63,6 @@ describe('Parser: apiBody', function () {
         optional: true,
         field: '\\MyClass\\field.user_first-name',
         defaultValue: 'John Doe',
-        description: 'Some description.',
-      },
-    },
-    {
-      title: 'All options, without optional-marker',
-      content: ' ( MyGroup ) { \\Object\\String.uni-code_char[] { 1..10 } = \'abc\', \'def\' }  ' +
-                     '\\MyClass\\field.user_first-name = \'John Doe\' Some description.',
-      expected: {
-        group: 'MyGroup',
-        type: '\\Object\\String.uni-code_char[]',
-        size: '1..10',
-        allowedValues: ['\'abc\'', '\'def\''],
-        optional: false,
-        field: '\\MyClass\\field.user_first-name',
-        defaultValue: 'John Doe',
-        description: 'Some description.',
-      },
-    },
-    {
-      title: 'All options, without optional-marker, without default value quotes',
-      content: ' ( MyGroup ) { \\Object\\String.uni-code_char[] { 1..10 } = \'abc\', \'def\' }  ' +
-                     '\\MyClass\\field.user_first-name = John_Doe Some description.',
-      expected: {
-        group: 'MyGroup',
-        type: '\\Object\\String.uni-code_char[]',
-        size: '1..10',
-        allowedValues: ['\'abc\'', '\'def\''],
-        optional: false,
-        field: '\\MyClass\\field.user_first-name',
-        defaultValue: 'John_Doe',
         description: 'Some description.',
       },
     },
